Huge rewrite of the stats to go along with the last commit. Got rid of IStat<>, which is more important than it sounds. Before, every stat would create one class, which would house the stat's value, Changed event, and the stat type. This commit allows us to get rid of that object reference (8-or-so bytes) and the event (another 8-or-so bytes). Even better, the IStatCollection<>s had to listen to the Changed event of every stat in the collection (another about 4 bytes, or 8 on x64, for delegate reference). This all created a huge amount of memory overhead, and a bit of performance overhead. Now, each stat takes up only as much memory as is needed for the value type (2 bytes default, since its a short) and the StatType (a byte). Pretty important seeing as every Character and Item has 2 of every stat (base and mod, or base and req), and plenty of other things hold stat collections (status effects, skills, etc).

+++ netgore/trunk/DemoGame.Server/World/Map/Entities/Characters/NPCChat/ResponseActions/RestoreHP.cs
@@ -20,7 +20,7 @@ namespace DemoGame.Server.NPCChat.ResponseActions
         /// <param name="npc">The npc.</param>
         protected override void DoExecute(User user, NPC npc)
         {
-            user.HP = user.ModStats[StatType.MaxHP];
+            user.HP = (int)user.ModStats[StatType.MaxHP];
         }
     }
 }
